requirements.txt
tensorboard==2.0.2
tensorflow-datasets==3.2.1
tensorflow-gpu==2.0.1
Pillow==7.2.0
numpy==1.18.1